overhaul fixture and test structure
