+++ src/sds_data_model/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Test suite for the sds_data_model package."""
