+++ src/sds_data_model/tests/data/larger_cell_size_same_shape.tif
