+++ src/sds_data_model/tests/data/same_cell_size_same_shape.tif
