+++ src/sds_data_model/tests/fixtures/fixtures_DataFrameWrapper.py
@@ -1,4 +1,4 @@
-"""Tests for DataFrame wrapper class."""
+"""Tests for Metadata module."""
 from dataclasses import asdict
 from pathlib import Path
 from typing import Any, Dict, Iterable, Optional, Sequence, Union
@@ -99,26 +99,6 @@ def expected_empty_metadata() -> None:
     return None
 
 
-def test_vector_layer_from_files(
-    spark_session: SparkSession,
-    temp_path: str,
-    expected_name: str,
-    expected_dataframe: SparkDataFrame,
-    expected_metadata: None,
-) -> None:
-    """Reading test data returns a DataFrameWrapper with expected values."""
-    received = DataFrameWrapper.from_files(
-        spark=spark_session,
-        data_path=temp_path,
-        read_file_kwargs={"header": True, "inferSchema": True},
-        name="Trial csv",
-    )
-
-    assert received.name == expected_name
-    assert received.metadata == expected_empty_metadata
-    assert_df_equality(received.data, expected_dataframe)
-
-
 @fixture
 def expected_dataframe_limit(
     spark_session: SparkSession,
@@ -176,63 +156,7 @@ def expected_dataframe_filter(
         data=data,
         schema=expected_schema,
     )
-
-
-@pytest.mark.parametrize(
-    argnames=(
-        "method_name",
-        "method_args",
-        "method_kwargs",
-        "expected_dataframe_name",
-    ),
-    argvalues=(
-        ("limit", None, {"num": 2}, "expected_dataframe_limit"),
-        ("select", ["a", "b"], None, "expected_dataframe_select"),
-        ("filter", "a = 3", None, "expected_dataframe_filter"),
-    ),
-    ids=(
-        "limit",
-        "select",
-        "filter",
-    ),
-)
-def test_call_method(
-    spark_session: SparkSession,
-    temp_path: str,
-    method_name: str,
-    method_args: Optional[Union[str, Sequence[str]]],
-    method_kwargs: Optional[Dict[str, Any]],
-    expected_dataframe_name: str,
-    request: FixtureRequest,
-) -> None:
-    """Function to test most common methods used by call_method."""
-    received = DataFrameWrapper.from_files(
-        spark=spark_session,
-        data_path=temp_path,
-        read_file_kwargs={"header": True, "inferSchema": True},
-        name="Trial csv",
-    )
-
-    if method_args and method_kwargs:
-        # If we get both, use both, unpacking the `method_kwargs` dictionary.
-        received.call_method(method_name, method_args, **method_kwargs)
-    elif method_args and isinstance(method_args, list):
-        # If we get only method_args, and it's a `list`, use it.
-        received.call_method(method_name, *method_args)
-    elif method_args:
-        # Now method_args has to be a single item, use it.
-        received.call_method(method_name, method_args)
-    elif method_kwargs:
-        # And `method_kwargs` has to be a dictionary, so unpack it.
-        received.call_method(method_name, **method_kwargs)
-    else:
-        # Do nothing
-        pass
-
-    expected_dataframe = request.getfixturevalue(expected_dataframe_name)
-    assert_df_equality(received.data, expected_dataframe)
-
-
+    
 @fixture
 def schema_other() -> StructType:
     """Schema of `other` DataFrame for joining."""
@@ -291,25 +215,7 @@ def expected_dataframe_joined(
         data=data,
         schema=expected_schema_joined,
     )
-
-
-def test_call_method_join(
-    spark_session: SparkSession,
-    temp_path: str,
-    dataframe_other: SparkDataFrame,
-    expected_dataframe_joined: SparkDataFrame,
-) -> None:
-    """Passing the `.join` method to `.call_method` produces the expected results."""
-    received = DataFrameWrapper.from_files(
-        spark=spark_session,
-        data_path=temp_path,
-        read_file_kwargs={"header": True, "inferSchema": True},
-        name="Trial csv",
-    )
-    received.call_method("join", other=dataframe_other, on="a")  # type: ignore[arg-type]  # noqa: B950
-    assert_df_equality(received.data, expected_dataframe_joined)
-
-
+    
 @fixture
 def hl_schema() -> StructType:
     """Schema HL DataFrame."""
@@ -401,41 +307,3 @@ def expected_hl_dataset() -> Dataset:
         coords=coords,
         attrs=expected_attrs,
     )
-
-
-def test_to_zarr(
-    hl_zarr_path: str,
-    expected_hl_dataset: Dataset,
-) -> None:
-    """The HL DataFrame wrapper is rasterised as expected."""
-    hl_dataset = open_dataset(
-        hl_zarr_path,
-        engine="zarr",
-        decode_coords=True,
-        mask_and_scale=False,
-        chunks={
-            "eastings": 10_000,
-            "northings": 10_000,
-        },
-    )
-
-    assert_identical(hl_dataset, expected_hl_dataset)
-
-
-def test_to_zarr_attrs(
-    hl_zarr_path: str,
-    expected_hl_dataset: Dataset,
-) -> None:
-    """The HL DataFrame wrapper is rasterised as expected."""
-    hl_dataset = open_dataset(
-        hl_zarr_path,
-        engine="zarr",
-        decode_coords=True,
-        mask_and_scale=False,
-        chunks={
-            "eastings": 10_000,
-            "northings": 10_000,
-        },
-    )
-
-    assert_identical(hl_dataset.attrs, expected_hl_dataset.attrs)
