+++ src/sds_data_model/tests/fixtures/fixtures_metadata.py
@@ -1,9 +1,30 @@
 """Tests for Metadata module."""
+
+from dataclasses import asdict
 from pathlib import Path
+from typing import Any, Dict, Iterable, Optional, Sequence, Union
 
 import pytest
-from pytest import fixture
+from chispa.dataframe_comparer import assert_df_equality
+from dask.array import concatenate, ones, zeros
+from numpy import arange
+from pyspark.sql import DataFrame as SparkDataFrame
+from pyspark.sql import SparkSession
+from pyspark.sql.types import (
+    ArrayType,
+    BinaryType,
+    IntegerType,
+    StringType,
+    StructField,
+    StructType,
+)
+from pytest import FixtureRequest, fixture
+from shapely.geometry import box
+from xarray import DataArray, Dataset, open_dataset
+from xarray.testing import assert_identical
 
+from sds_data_model.constants import BNG_XMAX, BNG_XMIN, BNG_YMAX, BNG_YMIN, CELL_SIZE
+from sds_data_model.dataframe import DataFrameWrapper
 from sds_data_model.metadata import Metadata
