+++ src/sds_data_model/tests/test_graph.py
@@ -9,14 +9,6 @@ from pytest import fixture
 from sds_data_model.vector import VectorLayer
 
 
-@fixture
-def expected_dag(shared_datadir: Path) -> str:
-    """The DAG we expect our function to generate."""
-    path_to_dot_file = shared_datadir / "ramsar_dag.dot"
-    dag: str = Source.from_file(path_to_dot_file).source
-    return dag
-
-
 def test_graph(shared_datadir: Path, expected_dag: Digraph) -> None:
     """Reads data, then converting to tiles, produces the expected graph."""
     data_path = str(shared_datadir / "Ramsar__England__.zip")
