+++ src/sds_data_model/tests/test_vector.py
@@ -13,47 +13,7 @@ from sds_data_model.constants import BBOXES, CategoryLookups, Schema
 from sds_data_model.vector import TiledVectorLayer, VectorLayer
 
 
-@fixture
-def expected_name() -> str:
-    """Expected VectorLayer name."""
-    return "ctry_21_gb_buc"
 
-
-@fixture
-def expected_metadata() -> None:
-    """Expected VectorLayer metadata."""
-    return None
-
-
-@fixture
-def expected_schema() -> Schema:
-    """Expected VectorLayer schema."""
-    return {
-        "OBJECTID": "int32",
-        "CTRY21CD": "object",
-        "CTRY21NM": "int8",
-        "CTRY21NMW": "object",
-        "BNG_E": "int32",
-        "BNG_N": "int32",
-        "LONG": "float64",
-        "LAT": "float64",
-        "GlobalID": "object",
-        "SHAPE_Leng": "float64",
-        "SHAPE_Area": "float64",
-    }
-
-
-@fixture
-def expected_category_lookups() -> CategoryLookups:
-    """Expected VectorLayer category lookups."""
-    return {
-        "CTRY21NM": {
-            -1: "No data",
-            0: "England",
-            1: "Scotland",
-            2: "Wales",
-        },
-    }
 
 
 def test_vector_layer_from_files(
@@ -100,34 +60,7 @@ def test_tiled_vector_layer_from_files(
     assert received.category_lookups == expected_category_lookups
 
 
-@fixture
-def gpdf() -> GeoDataFrame:
-    """Dummy GeoDataFrame."""
-    return GeoDataFrame(
-        data={
-            "category": ["A", "B", "C", None],
-            "geometry": [box(*bbox) for bbox in BBOXES[0:4]],
-        },
-        crs="EPSG:27700",
-    )
 
-
-@fixture
-def expected_HL_array() -> NDArray[int8]:
-    """The expected array for grid cell HL."""
-    return zeros(
-        shape=(10_000, 10_000),
-        dtype=dtype("int8"),
-    )
-
-
-@fixture
-def expected_HM_array() -> NDArray[int8]:
-    """The expected array for grid cell HM."""
-    return ones(
-        shape=(10_000, 10_000),
-        dtype=dtype("int8"),
-    )
 
 
 def test_pipeline(
