--- tests/__init__.py
@@ -1,1 +1,0 @@
-"""Test suite for the sds_data_model package."""
