--- tests/conftest.py
@@ -1,8 +1,0 @@
-"""Central place for test fixtures."""
-from fixtures.fixtures_metadata import (
-    expected_metadata,
-    expected_without_keywords,
-    expected_without_title,
-    local_file_path,
-    remote_url,
-)
